Version 4 of Op-Amp library
get_resistors function completed - returns a pointer the the address of the first value of an array of the resistors. the subsequent resistors can be found but calling the following address from the first. Base simulation functionality has been added, still some kinks to be worked out in the use-case.