Version 5 of Op-Amp library
Further developed simulation functionality to take input variables for the wave-form, and store in an array. Currently the programme is being converted to store in a vector instead in preparation for matplot++ which plots from vectors. The function returns a pointer to the address of the first value of the array. Subsequently, the following 511 values can be found by adding 1 to this address in hex. Currently after a certain point the values stored in the addresses are not as they should be (they are either very close to zero or very large values). This is an error to be fixed in version 6.