Final Main Version of the op_amp.h
Added file handling functionality to write the simulation data to a csv file to then be plotted in excel or with a python script. The only remaining potential change in version 7.1 will be to add a try catch for if the csv file already exists - Currently the csv file could be overwritten or potential undefined behaviour may occur.